use map-reduce method for all priorities method
